+++ src/core/shared/flags/StorageRules.ts
@@ -1,0 +1,9 @@
+/**
+ * An enum that defines storage checks / behavior propagation
+ * OPEN
+ * @export
+ * @enum {number}
+ */
+export const enum STORAGE_RULES {
+    OPEN = 'storage-open',
+}